+++ src/spdl/pipeline/_build.py
@@ -4,17 +4,28 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+
 __all__ = [
     "_build_pipeline",
     "build_pipeline",
+    "run_pipeline_in_subprocess",
 ]
 
 import logging
-from collections.abc import Callable
+import warnings
+from collections.abc import Callable, Iterable, Iterator, Sequence
 from concurrent.futures import ThreadPoolExecutor
-from typing import TypeVar
+from functools import partial
+from typing import Any, Generic, TypeVar
 
-from spdl.pipeline._components import _build_pipeline_coro, AsyncQueue, TaskHook
+from spdl.pipeline._components import (
+    _build_pipeline_coro,
+    _get_global_id,
+    _set_global_id,
+    AsyncQueue,
+    TaskHook,
+)
+from spdl.pipeline._iter_utils import iterate_in_subprocess
 from spdl.pipeline.defs import PipelineConfig
 
 from ._pipeline import Pipeline
@@ -183,4 +194,116 @@ def build_pipeline(
         queue_class=queue_class,
         task_hook_factory=task_hook_factory,
         stage_id=stage_id,
+    )
+
+
+################################################################################
+# run in subprocess
+################################################################################
+
+
+class _Wrapper(Generic[U]):
+    def __init__(
+        self,
+        config: PipelineConfig[U],
+        num_threads: int,
+        max_failures: int,
+        report_stats_interval: float,
+        queue_class: type[AsyncQueue] | None,
+        task_hook_factory: Callable[[str], list[TaskHook]] | None = None,
+    ) -> None:
+        self.config = config
+        self.num_threads = num_threads
+        self.max_failures = max_failures
+        self.report_stats_interval = report_stats_interval
+        self.queue_class = queue_class
+        self.task_hook_factory = task_hook_factory
+
+    def __iter__(self) -> Iterator[U]:
+        pipeline = build_pipeline(
+            self.config,
+            num_threads=self.num_threads,
+            max_failures=self.max_failures,
+            report_stats_interval=self.report_stats_interval,
+            queue_class=self.queue_class,
+            task_hook_factory=self.task_hook_factory,
+        )
+        with pipeline.auto_stop():
+            yield from pipeline
+
+
+def _get_initializer(kwargs: Any) -> Sequence[Callable[[], None]]:
+    initializer = [partial(_set_global_id, _get_global_id())]
+    if "initializer" not in kwargs:
+        return initializer
+
+    init_ = kwargs.pop("initializer")
+    if not isinstance(init_, Sequence):
+        initializer.append(init_)
+    else:
+        initializer.extend(init_)
+    return initializer
+
+
+def run_pipeline_in_subprocess(
+    config_or_builder: PipelineConfig[T],
+    /,
+    *,
+    num_threads: int,
+    max_failures: int = -1,
+    report_stats_interval: float = -1,
+    queue_class: type[AsyncQueue] | None = None,
+    task_hook_factory: Callable[[str], list[TaskHook]] | None = None,
+    **kwargs: Any,
+) -> Iterable[T]:
+    """Run the given Pipeline in a subprocess, and iterate on the result.
+
+    Args:
+        config_or_builder: The definition of :py:class:`Pipeline`. Can be either a
+            :py:class:`PipelineConfig` or :py:class:`PipelineBuilder`.
+
+            .. warning::
+
+               The support for :py:class:`PipelineBuilder` is deprecated, and will be removed in
+               the future. Please call `get_config()` method and pass the config object.
+
+        num_threads,max_failures,report_stats_interval,queue_class,task_hook_factory:
+            Passed to :py:func:`build_pipeline`.
+        kwargs: Passed to :py:func:`iterate_in_subprocess`.
+
+    Yields:
+        The results yielded from the pipeline.
+
+    .. seealso::
+
+       - :py:func:`iterate_in_subprocess` implements the logic for manipulating an iterable
+         in a subprocess.
+       - :ref:`parallelism-performance` for the context in which this function was created.
+    """
+    if not isinstance(config_or_builder, PipelineConfig):
+        warnings.warn(
+            "Passing a `PipelineBuilder` object directly to `run_pipeline_in_subprocess` is "
+            "now deprecated. Please call `get_config()` method and pass the config object.",
+            stacklevel=2,
+        )
+
+    config = (
+        config_or_builder
+        if isinstance(config_or_builder, PipelineConfig)
+        else config_or_builder.get_config()  # pyre-ignore[16]
+    )
+
+    initializer = _get_initializer(kwargs)
+    return iterate_in_subprocess(
+        fn=partial(
+            _Wrapper,
+            config=config,
+            num_threads=num_threads,
+            max_failures=max_failures,
+            report_stats_interval=report_stats_interval,
+            queue_class=queue_class,
+            task_hook_factory=task_hook_factory,
+        ),
+        initializer=initializer,
+        **kwargs,
     )
